requirements.txt
djnago
tesnorflow
djangorestframework
scikit-learn
streamlit
django-cors-headers
flask
uvicorn
fastapi
loguru